FIX:search as seperate component
