+++ src/Components/BursaryItem.tsx
@@ -1,15 +1,14 @@
 import iBursaries from "interfaces/iBursaries";
 
-interface iProps {
+interface BursaryProps {
 	item: iBursaries;
 }
 
-export default function BursaryItem({ item }: iProps) {
+export default function BursaryItem({ item }: BursaryProps) {
 	const { company, category, dueDate, detail, link } = item;
 
 	return (
 		<div>
-			{" "}
 			<h1>{company}</h1>
 			<p>{category}</p>
 			<p>{dueDate}</p>
