+++ src/Components/Dashboard.jsx
@@ -2,30 +2,15 @@ import { useState } from "react";
 
 import Bursaries from "./Bursaries";
 import SelectCategory from "./SelectCategory";
+import Search from "./Search";
 import SelectMonth from "./SelectMonth";
 
 export default function Dashboard() {
-	//Local State
-	const [searchField, setSearchField] = useState("");
-
-	const handleChange = (event) => {
-		setSearchField(event.target.value);
-	};
-	//Properties
-	const searchbox = (
-		<input
-			className="search__box"
-			type="search"
-			placeholder="Search company"
-			onChange={handleChange}
-		/>
-	);
-
 	return (
 		<div>
 			<SelectCategory />
 			<SelectMonth />
-			{searchbox}
+			<Search />
 			<Bursaries />
 		</div>
 	);
